fix: move context to @lib/context.tsx
