+++ lib/wagmi.ts
@@ -1,4 +1,3 @@
-import { cookieStorage, createStorage } from '@wagmi/core'
 import { WagmiAdapter } from '@reown/appkit-adapter-wagmi'
 import { mainnet, arbitrum } from '@reown/appkit/networks'
 
@@ -13,9 +12,6 @@ export const networks = [mainnet, arbitrum]
 
 //Set up the Wagmi Adapter (Config)
 export const wagmiAdapter = new WagmiAdapter({
-  storage: createStorage({
-    storage: cookieStorage
-  }),
   ssr: true,
   projectId,
   networks
